Set up CI (#15)
This PR configures CI with some rake tasks, so that we can move forward
more confidently while collaborating. I also fixed existing violations.

Note that for the Rust code where we're dereferencing pointers, clippy
forces us to mark the function as unsafe and to document why it is
unsafe. For now, I just went ahead with a generic comment since I'm not
sure we'll even need to keep these functions around.