Update ID structs to use Hash directly instead of byte value (#76)
Using the byte value of blake3 hashes is a bit tricky because it becomes
tricky to encode the byte values in to strings for printing and DB
storage. Passing around the `blake3::Hash` struct instead will allow us
to take advantage of the methods on `blake3::Hash` for equality
comparison and conversion to hex strings. This should make DB storage,
printing, and comparison easy.

Upon comparing the size of a `blake3::Hash` instance and the byte
encoded hash value, both take up 32 bytes -- so we shouldn't be
incurring any extra memory consumption.