Make the index a graph (#69)
Closes #11

Make the index implementation a graph structure, supporting synchronizing newly indexed definitions. There are a few changes in this PR:

1. Now that the Index is all implemented in a top level graph, with no sub-graphs inside of nodes, merging different `Index` instances becomes much simpler. I made each `RubyIndexer` create its own local index and we merge everything at the end
2. I got rid of the `Declaration` concept, at least for now. Right now, we have global fully qualified names that have relationships with all definitions that compose that name. This effectively moves the sub-graph we had inside of the `Declaration` struct into the top level graph - which is the key change in this PR

Here's a visual representation of what the graph looks like and where each part is stored.

<img width="2659" height="1198" alt="image" src="https://github.com/user-attachments/assets/4c042991-0562-4951-8564-8f01d6ef56ed" />

As we advance in the implementation, we are likely going to add more edges (relationships) to the graph. For example, inheritance can be represented with new inheritance edges between name IDs.