Add memory leak CI check (#113)
Closes #107 

This PR adds a CI check for verify for memory leaks in both the Rust
part of the code and the C layer that connects it to the Ruby VM.

Both levels of checks work similarly. They add runtime checks to the
code so that if we allocate any memory and forget to free it, they can
report that as a problem and let us know.

Rust offers [many types of
sanitizers](https://doc.rust-lang.org/beta/unstable-book/compiler-flags/sanitizer.html)
to check for problems and for the C layer we're using
[ruby_memcheck](https://github.com/Shopify/ruby_memcheck) (which uses
valgrind under the hood).

The CI check will:

1. Compile the crate with some form of sanitization
2. Run our Ruby test suite with ruby_memcheck
3. Report any issues

I added 3 Rust sanitizers that may be of interest to us:

- Leak checks for memory leaks
- Memory checks for uninitialized reads
- Thread checks search for data races