Use a blob column to store definition specific data (#135)
Each definition may have its own specific fields that don't match other
definitions. To account for this, what Pyre does is putting the
definition specific data into a serialized blob column and only exposing
columns it needs for querying.

I believe we should take the same approach. With the addition of #125,
it's no longer possible to correctly dump the graph to the database
because methods have extra fields other definitions do not have.

This PR changes the start and end offset columns to be a generic data
blob column, which we store MessagePack serialized data for the
definition in.

This is required to unblock #130.