Add ability to load all data related to a URI (#130)
Allow loading all graph data related to a URI from the database.

In addition to adding the loading logic, I also made the connection
persistent. The previous approach was really annoying because in-memory
databases are destroyed immediately if there are no connections, which
meant that we would need to over-expose the API for testing.

I think this approach will also make it easier to handle failures, like
the user deleting or corrupting the database file - which we can recover
from.