Remove Graph arc and use mutable reference (#148)
This PR proposes removing the need for the graph to always be inside an
`Arc<Mutex>`. This makes most of the code more convenient, but there's
one caveat.

With `Arc`, when we're crossing the boundary from the Ruby VM into the
Rust crate, we can very easily increment the Arc strong count and ensure
the graph is not accidentally freed prematurely (leaving a dangling
pointer stored in the Ruby VM's object).

To remove the `Arc`, we need to use a `Box`. The issue is that boxes
don't count references, so we need to manually forget the box every
single time we turn the raw pointer into a `Graph`, otherwise Rust will
free it and the VM is left with a dangling pointer.

That can get tedious and error prone very fast, so I tried implementing
a `with_graph` function that accepts a closure, which removes the need
to manually have to forget the graph object. It's not the prettiest
code, but at least the surface area is limited to the Rust to C
interface.

Let me know what you think!