Sync with DB during indexing
The current indexing flow, when we call the indexing entrypoint, does
not interact with the db. We have the api ready but it's not hooked into
the indexing flow yet.

Current flow:

indexing starts -> collect documents from workspace -> index

(no db interactions)

Updated flow:

indexing starts -> collect documents from workspace -> diff against db
state -> delete stale entries from db -> index -> add new entries to db

To be able to know if the file has changed (and thus needs re-indexing)
we will introduce a content_hash to the documents table in the db. Then
during the diff checking stage we will

First compare the uris for added/deleted files

The added files will collected for indexing

The deleted files will be removed from the db (with their corresponding
entities)

Then for the existing files, compare the content_hash to collect the
files that need reindexing

When we partially delete/create entries we shouldn't need to worry about
cascading updates to entries from documents that remain unchanged,
because we have not implemented more complex relationships in the graph
yet (i.e., inheritance). The relationships that we do have now are
already taken care of by the database constraints.