Use i64 instead of u64 for ID
This is to fix the issue where the db Integer field expects an i64
value. We are currently trying to pass in a u64 value (as string) for
which a type mismatch error would be raised when we try to save the
data.

So here we cast the u64 hash result into i64 to meet the db constraint.