Introduce `DeclarationId` (#213)
Trying to figure out the resolution phase, @jenny-codes, @Morriar and me
figured out that we were missing one type of connection in our graph -
that both Sorbet and Pyre have.

We currently don't have a connection between a `Declaration` and
everything that it owns through an unqualified name. This makes it
impossible to resolve constant references (the reason will be clearer in
the next few PRs).

After a bit of studying we concluded that we actually need IDs for BOTH
fully qualified names and unqualified names, to save memory on storing
strings while representing all of the concepts we need.

This PR doesn't actually implement anything, it just renames `NameId` to
`DeclarationId` (the ID of the global representation of an entity) and
also _keeps_ `NameId` for unqualified references. The rename touches a
lot of files, so I wanted to create a separate PR just for this.

Here's an example of what it will look like:

```ruby
module Foo    # DeclarationId(Foo), NameId(Foo)
  class Bar   # DeclarationId(Foo::Bar), NameId(Bar)
    CONST = 1 # DeclarationId(Foo::Bar::CONST), NameId(CONST)
  end
end
```