Add members mapping to declarations (#218)
Another step towards constant reference resolution. This PR adds the
`members` mapping to the global declarations, so that we can keep track
of what is "owned" by each namespace.

For example

```ruby
module Foo
  class Bar
    def baz; end
  end
end
```

In this case, `Foo` is a member of `Object` (which owns all top level
constants), `Bar` is a member of `Foo` and `baz` is a member of `Bar`.

The connection is made with `NameId` (unqualified name) to
`DeclarationId` (fully qualified name). This allows us to traverse the
graph checking if something is defined at each step - which enables
resolution.

**Note**: this is currently not being used. The reason is that we need
more changes to the indexer to track a stack of owners. We need to
remember which namespace we're currently inside to add new definitions
as members. We don't currently track that, so I will follow up with a PR
to do so, which allows us to populate the `members` map.