Add content_hash to document struct (#209)
For https://github.com/Shopify/index/issues/88
Breaking up https://github.com/Shopify/index/pull/173 into smaller chunks.

Here we add content_hash to the Document model.

We are choosing to use `u16`​ here for the content_hash as the purpose of this field is for staleness check between two versions of a document, which is a different use case from the IDs that need the full 64 bits.

Collision risk comparison:

(how likely we would fail to detect a change in the document content for the number of files we have)

| Hash Size | Values | 100 Files Collision Risk |
| --- | --- | --- |
| u8 | 256 | 32\.4% |
| u16 | 65,536 | 0\.15% |
| u32 | 4\.3B | 0\.000002% |

For 100 file updates the collision risk is 0.15% which feels acceptable.