Avoid advancing the nesting if not necessary (#241)
This PR stops advancing our nesting stacks for constructs that aren't
modules or classes (namespaces). This PR is just a small refactor that
makes the implementation of the parent PR in this Graphite stack easier
to review.

Basically, I added a `fully_qualify` method to produce a fully qualified
name without doing stack manipulation and I also made some minor
modifications in how they are advanced.

**Warning**: it's not SUPER relevant to get too attached to the
implementation because the parent PR changes it significantly. Prefer
reviewing the parent!