Make result local graph optional (#226)
Currently when an error happens during file reading we will initialize a
RubyIndexer object, store the file reading errors, skip the indexing,
and return the empty local graph and the list of errors.

We can simplify it by making the local graph an optional return value,
and only return it when there is no error from the file reading step.
This way we save an RubyIndexer initialization as well as an attempt to
merge an empty local graph back to the global one.

This is also paving the way for the incremental sync which will start
skipping the indexing of existing files whose content have not changed.