Populate declaration members (#249)
Start populating the members of each declaration. Being able to inspect
which members exist in each global declaration allows us to perform
constant resolution, but also provide completion based on what's
available within the current context.

This is still missing the logic to handle deletions, but with this we're
able to implement resolving constants based on lexical scopes.

### Implementation

An important aspect of the implementation that I had to test out is what
things are associated to when they are defined in the top level. There
are essentially two cases:

1. Constants of any kind (modules, classes and other constants) are
associated to `Object` when defined at the top level
2. Anything else (instance variables, methods and so on) get associated
to `<main>`. Ruby uses a special instance of the `Object` class to
represent its top level. Performing some tests on a script, you can get
a better sense for the behaviour

I started adding the `<main>`​ declaration artificially since it doesn't
actually exist anywhere, it's created by Ruby
[here](https://github.com/ruby/ruby/blob/9ad902e55c610e66114f528f77f7895295a242de/vm.c#L4693).
This is equivalent to Sorbet's `<root>`, but I figured why not call it
the same as Ruby does.

```ruby
attr_reader :foo # => crash! <main> is an instance of `Object` and therefore attr_* methods don't exist
@@class_var = 123 # => crash! can't use class variables on <main>

def foo; end

foo() # ok! `foo` was defined in the <main> instance

obj = Object.new
obj.foo # crash! `foo` is not defined in the `Object` class
```