Add owner_id relationship to declarations (#268)
We're now indexing the `members` information, but we have no way of
removing entries from it when someone deletes an existing definition. To
be able to work backwards from the definition to the members that must
be cleaned, we need the reverse of the `members` relationship.

This PR introduces the `owner_id` to declarations, so that we can walk
the graph in the opposite direction as well. I'll follow up with the PR
to delete `members` entries when definitions are removed, which should
be the last piece in managing the `members` data.