Remove database temporarily (#301)
After a few discussions, the team agreed to remove all database related
implementations temporarily. This PR removes all DB related things
including dependencies, SQL files and so on.

### Context on the decision

While talking about the database, we increasingly realized that there
are several use cases and scenarios we haven't really thought out in
complete detail:

- Can all tools that use Saturn share the same database cache? What if
they only care about of portion of the data (think not caring about RBI
files or wanting to index ERB)?
- Should all tools ensure the database is always up to date with the
most complete information possible and then use only what they need?
What if they want to store metadata that other tools don't care about?

Another question that came up was: is the database more important from a
CPU perspective (to cache and avoid repeated work) or from a memory
perspective (to offload what's not being used to the disk)? And are the
two perspectives equally relevant?

Considering that:

- The database is a performance optimization, but our current
performance is already significantly better than the Ruby LSP and Spoom
implementations
- We have all of these questions to dig further into
- Simply getting the functionality working properly is already a LOT of
work

it makes more sense to finalize the functionality end to end first and
leave the database for later. That way, we can ensure that we have a
working product and then evaluate how to plug the database considering
every angle that this project has to support.

We will bring the database and caching back later, but for now removing
it will allow us to move faster, with fewer dependencies.