Split constant and method references into their own maps (#303)
It was quite odd to have to deal with any type of reference everywhere. We mostly just added `unreacheable` or something of the sort to appease the type system when we knew something wasn't possible.

That's not super nice, so I propose we split each type of reference into their own map. That's all this PR does, splitting method and constant references into distinct hashmaps.

### Notes

I experimented with creating different ID types for constants and methods. I think it's nice, but the moment we do it, then the question becomes: what happens to the list of resolved references inside declarations?

We could put all types of reference IDs in an enum, but that's a bit dirty. It's not valid to have a `MethodRefId` inside a `ClassDeclaration` for example.

If we want to have distinct IDs for references, I'd suggest we also specialize `Declaration` into an enum for all different types of entities that exist.

This is why I did not include this as part of the PR.

### Performance

No difference when it comes to indexing time, but a bit of a reduction in memory usage. In Core:

```
BEFORE
2218.73 MB
AFTER
1633.59MB
```