Start tracking parent classes (#304)
This PR starts tracking parent classes by assigning a constant reference ID to the `ClassDefinition` struct.

Our idea is that the definition can remember the ID without actually knowing whether the reference has already been resolved or not. That allows us to save a bit of work from not having to resolve a reference again if it was already resolved before and not changed.