Recursively resolve constant parent scopes (#312)
Start recursing on constant parent scopes to resolve each part.

For now, I just hacked something that works for the `resolve_references`​ method, but I'm going to revisit this in more detail for designing the resolution phase. We're in fact doing more work than necessary because when we recurse on parent scopes, we don't remember that we already resolve it anywhere - which means we always resolve them multiple times. Anyway, the will be addressed in #200.