Use line-index crate for offset to position conversions (#325)
In #126, I made a bit of a faux pas. I totally forgot to call attention
to the fact that Rust Analyzer already publishes a number of crates for
LSP related things (such as translating offsets into line/col
information) and that we should check what's available.

Indeed, they do publish the `line-index` crate, which does everything we
need to translate locations including all of the different encodings.

I propose that we switch to using the crate so that we can avoid having
to maintain this translation. The crate supports translating in both
directions and it supports handling incremental edits as well.

In the future, we can make each `Document` instance hold their own
`LineIndex` instance and we'll be able to easily translate positions in
both directions - regardless of whether the document has been committed
to disk or not.