Make class and module names constant references (#320)
This PR makes classes and modules store a constant reference ID rather
than a name. This allows us to delay naming classes and modules to the
resolution phase, when we can actually figure out the correct fully
qualified name.



For now, I didn't make any modifications to resolution to minimize the
amount of changes in one go. So we're still incorrectly resolving more
complex cases. All we're doing is building the fully qualified name with
each nesting part of the name.



I will follow up with changes to make the naming happen during
resolution.