Re-enable Windows CI (#324)
Closes #68

It seems that https://github.com/ruby/prism/pull/3594 _did_ fix the
Windows build 🎉. This PR re-enables it so that we don't accidentally add
Windows incompatible code.

Most of the fixes are related to assuming forward slashes in file paths
or not accounting for the carriage return character where it may happen.