Ensure constant reference resolution results are cached (#322)
This PR ensures that intermediate resolution results are cached in the
graph (so that we don't try to resolve the same constant multiple
times).

Additionally, I moved all of the algorithm to `resolution.rs`. Passing a
reference to the graph will make it easier to add parallelism afterwards
since the graph can't put itself inside an `Arc`.