Compute all global information during resolution (#331)
Step towards #200

This PR starts creating all declarations with their memberships during
the resolution phase.

The idea is to sort all constant definitions and references based on the
complexity of their name, so that processing units in the beginning of
the list unlock work that is further ahead and so on.

I tried to document the code, but here are a few explanations.

### Name complexity

Each name part that appears in the nesting counts as 1 to compute the
complexity used to order the list. For example:

```ruby
module Foo # 1
  module Bar # 2
    module Baz # 3
    end
  end

  module Qux::Boop # 3
    module Zip # 4
    end
  end
end
```

I named this "complexity score", but please suggest other names if you
have ideas. This is simply a score to sort constants in the most likely
order of reducing resolution failures.

### Handling parent scopes in definitions

When we don't have a parent scope in a module/class/constant definition
(i.e.: a simple name), there's no need to run a resolution algorithm. We
can just take the name as is to create the declaration concept.

Thanks to the ordering based on complexity, we should also guarantee
that we will never be in a situation where we find a constant with a
parent scope or nesting that hasn't yet been resolved - at least until
ancestors are added in the mix.

For constants with parent scopes, we first always resolve the reference
to figure out what it points to before naming the definition.