Index variations of global variables (#357)
Closes #82

This PR adds indexing for the remaining types of global variables. To my
surprise, in all of these variant of global variable nodes, Ruby does
not crash if the variable doesn't exist beforehand - including the alias
global variable node!

None of these crash:

```ruby
$a &= 1
$a &&= 1
$a ||= 1
alias $a $b
```

Because of that, I modelled all of them as definitions, so that we don't
end up adding errors that don't match the runtime (same idea as `A ||=
1`).