Add parent class linearization (#339)
Add parent class linearization. I'm more interested in seeing how this fits into the general resolution algorithm at this point, but I think this PR also uncovers the need for specializing declarations.

For example, if we turn `Declaration`​ into an enum, we can always pick the first definition we find as the truth for error scenarios, which makes it easier to emit diagnostics. Imagine we find `class Foo`​ and `module Foo`​. When we're processing the second `Foo`​, it would be easier to recognize the mistake by looking at the `Declaration`​. Same thing for superclass mismatches.

In terms of the implications for the resolution phase, I used a `OnceCell`​for ancestors, which gives us interior mutability (avoiding borrow issues). This makes it easier to work with than `RefCell`​ because we never expect to mutate the ancestor vector. We always erase and set the entire vector in one go. By using `OnceCell`​, we don't pay the runtime price of `RefCell`​.